Lock scripts mem cache (#70)
* remove unnecessary argument

* cache script and control block

* use get_or_put to optimize cache usage

* remove profiling

* naming

* fix merge compile error

* assert commitment_public_keys is not empty

* ignoring NotFound error kind

* remove all old files if cache file count reach the limit

* address comments

* fix print on delete

+++ bridge/tests/bridge/integration/peg_out/disprove.rs
@@ -294,7 +294,7 @@ async fn test_disprove_success() {
     let verifier_reward_script = reward_address.script_pubkey(); // send reward to withdrawer address
     disprove.add_input_output(
         &config.connector_c,
-        script_index as u32,
+        script_index,
         disprove_witness,
         verifier_reward_script,
     );
